+++ index.d.ts
@@ -1,19 +1,67 @@
 type Tile = MBTiles.Tile
 type Metadata = MBTiles.Metadata
 
+/**
+ * MBTiles
+ */
 declare class MBTiles {
   constructor(uri: string)
+
+  /**
+   * Save buffer data to individual Tile
+   */
   save(tile: Tile, image: Buffer): Promise<boolean>
+
+  /**
+   * Retrieves Metadata from MBTiles
+   */
   metadata(): Promise<Metadata>
+
+  /**
+   * Delete individual Tile
+   */
   delete(tile: Tile): Promise<boolean>
+
+  /**
+   * Count the amount of Tiles
+   */
+  count(tiles?: Tile[]): Promise<boolean>
+
+
+  /**
+   * Update Metadata
+   */
   update(metadata: Metadata): Promise<Metadata>
+
+  /**
+   * Finds all Tile unique hashes
+   */
+  findAll(tiles?: Tile[]): Promise<Tile[]>
+
+  /**
+   * Finds one Tile and returns Buffer
+   */
   findOne(tile: Tile): Promise<Buffer>
-  findAll(): Promise<Tile[]>
+
+  /**
+   * Build SQL tables
+   */
   tables(): Promise<boolean>
-  count(): Promise<boolean>
+
+  /**
+   * Build SQL index
+   */
   index(): Promise<boolean>
+
+  /**
+   * Creates hash from a single Tile
+   */
   hash(tile: Tile): number
-  hashes(): Promise<MBTiles.Hashes>
+
+  /**
+   * Creates a hash table for all tiles
+   */
+  hashes(tiles?: Tile[]): Promise<MBTiles.Hashes>
 }
 
 declare namespace MBTiles {